- Added dealLink row to Post entity, Created CommentDto, Added validation, fixed delete and update

+++ backend/src/main/java/com/pepper/backend/service/CommentService.java
@@ -1,5 +1,7 @@
 package com.pepper.backend.service;
 
+import com.pepper.backend.dto.CommentDto;
+import com.pepper.backend.dto.MessageDto;
 import com.pepper.backend.dto.UpdateCommentDto;
 import com.pepper.backend.dto.CreateCommentDto;
 import com.pepper.backend.exception.ResourceNotFoundException;
@@ -8,12 +10,17 @@ import com.pepper.backend.model.User;
 import com.pepper.backend.repository.CommentRepository;
 import com.pepper.backend.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
+import java.security.Principal;
 import java.util.List;
+import java.util.stream.Collectors;
 
+@Transactional
 @Service
 @RequiredArgsConstructor
 public class CommentService {
@@ -38,10 +45,18 @@ public class CommentService {
         return commentRepository.save(comment);
     }
 
-    public void deleteComment(long id) {
+    public ResponseEntity<?> deleteComment(long id, Principal principal) {
+        User user = userRepository.findByUsername(principal.getName())
+                .orElseThrow(() -> new ResourceNotFoundException("Comment with id " + id + " not exist"));
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Comment with id " + id + " not exist"));
+
+        if (user.getUserId() != comment.getUserId()){
+            return ResponseEntity.ok(new MessageDto("You don't have permission to do it"));
+        }
+
         commentRepository.delete(comment);
+        return ResponseEntity.ok(new MessageDto("Comment successfully deleted"));
     }
 
     public Comment getCommentById(long id) {
@@ -49,14 +64,33 @@ public class CommentService {
                 .orElseThrow(() -> new ResourceNotFoundException("Comment with id " + id + " not exist"));
     }
 
-    public Comment updateCommentContent(long id, UpdateCommentDto UpdateCommentDto) {
+    public ResponseEntity<?> updateCommentContent(long id, UpdateCommentDto updateCommentDto, Principal principal) {
+        User user = userRepository.findByUsername(principal.getName())
+                .orElseThrow(() -> new ResourceNotFoundException("Comment with id " + id + " not exist"));
+
         Comment databaseComment = commentRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Comment with id " + id + " not exist"));
-        databaseComment.setContent(UpdateCommentDto.getContent());
-        return commentRepository.save(databaseComment);
+
+        if (user.getUserId() != databaseComment.getUserId()){
+            return ResponseEntity.ok(new MessageDto("You don't have permission to do it"));
+        }
+
+        databaseComment.setContent(updateCommentDto.getContent());
+        commentRepository.save(databaseComment);
+
+        return ResponseEntity.ok(new MessageDto("Comment successfully updated"));
     }
 
-    public List<Comment> getCommentsByPostId(Long postId){
-        return commentRepository.findByPostId(postId);
+    public List<CommentDto> getCommentsByPostId(Long postId){
+        List<Comment> comments = commentRepository.findByPostId(postId);
+
+        return comments.stream()
+                .map(comment -> ( new CommentDto(
+                        comment.getCommentId(),
+                        comment.getContent(),
+                        comment.getPostDate(),
+                        userRepository.findByUserId(comment.getUserId()).getUsername()))
+                )
+                .collect(Collectors.toList());
     }
 }
